+++ Sources/VaporCLI/cmd_clean.swift
